Add files via upload
Are you tired of being forgetfulness or lazy of doing any task? So i have a APP for you to help you to avoid forgetfulness or lazy by doing task.
This app will help you to remind anything that you forgot to do, will help you to remind whatever you will need or you will do. This app includes the following:
Assistant- Any.do Assistant takes care of your tasks, so you can spend time on the things that matter.
Calendar- Plan your day, week, month, and life.  Beautifully designed to help you rock your day, every day.
To-do-list- Keep your task, list and reminds in one place, always in-sync across all your devices.
We all need a little reminder to do things sometimes. We live in a world filled with distraction and various shiny things that often lead us off course of whatever it is we’re trying to do. Thus, setting reminders can not only be productive, but a savior in some situations. On Android, there are a ton of ways to create reminders and we took a look at the best reminder apps for android.